Replace Clojure functions that contain metadata with custom FnWithMeta (#150)
* Attempt at widening the arities for multimethod calls

There's a lot of `apply`/`RestFn`/`invoke` etc. dynamic call machinery
in Methodical's stack traces. This is an attempt to remove some of it
by going up to 7 direct args for multimethod calls. (And dispatch
functions.)

This hasn't removed much of the `apply` overhead in practice because
`with-meta` on a function wraps it with a naive function subclass that
always does a dynamic call. There are probably still some places that
more dynamic calls are creeping in, but I ran out of time to dig
deeper.

This may not go anywhere until I get back, but I wanted to publish this
just in case.

* Fix tests after widening arities

marg

* Introduce custom FnWithMeta to attach metadata to functions

* Reimplement threaded combinator with explicit first/last separation

---------

Co-authored-by: Braden Shepherdson <braden@metabase.com>